# docs/gateway/openapi.json
{
  "components": {
    "schemas": {
      "BudgetResponse": {
        "description": "Response model for budget information.",
        "properties": {
          "budget_duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Duration Sec"
          },
          "budget_id": {
            "title": "Budget Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "max_budget": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Budget"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "budget_id",
          "max_budget",
          "budget_duration_sec",
          "created_at",
          "updated_at"
        ],
        "title": "BudgetResponse",
        "type": "object"
      },
...
      "CreateBudgetRequest": {
        "description": "Request model for creating a new budget.",
        "properties": {
          "budget_duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Budget duration in seconds (e.g., 86400 for daily, 604800 for weekly)",
            "title": "Budget Duration Sec"
          },
          "max_budget": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Maximum spending limit",
            "title": "Max Budget"
          }
        },
        "title": "CreateBudgetRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PricingResponse": {
        "description": "Response model for model pricing.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "input_price_per_million": {
            "title": "Input Price Per Million",
            "type": "number"
          },
          "model_key": {
            "title": "Model Key",
            "type": "string"
          },
          "output_price_per_million": {
            "title": "Output Price Per Million",
            "type": "number"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "model_key",
          "input_price_per_million",
          "output_price_per_million",
          "created_at",
          "updated_at"
        ],
        "title": "PricingResponse",
        "type": "object"
      },
      "SetPricingRequest": {
        "description": "Request model for setting model pricing.",
        "properties": {
          "input_price_per_million": {
            "description": "Price per 1M input tokens",
            "title": "Input Price Per Million",
            "type": "number"
          },
          "model_key": {
            "description": "Model identifier in format 'provider:model'",
            "title": "Model Key",
            "type": "string"
          },
          "output_price_per_million": {
            "description": "Price per 1M output tokens",
            "title": "Output Price Per Million",
            "type": "number"
          }
        },
        "required": [
          "model_key",
          "input_price_per_million",
          "output_price_per_million"
        ],
        "title": "SetPricingRequest",
        "type": "object"
      },
      "UpdateBudgetRequest": {
        "description": "Request model for updating a budget.",
        "properties": {
          "budget_duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Duration Sec"
          },
          "max_budget": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Budget"
          }
        },
        "title": "UpdateBudgetRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "A clean FastAPI gateway for any-llm with API key management",
    "title": "any-llm Gateway",
    "version": "1.1.1.dev23+g56c5a9b20.d20251105"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/budgets": {
      "get": {
        "description": "List all budgets with pagination.",
        "operationId": "list_budgets_v1_budgets_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetResponse"
                  },
                  "title": "Response List Budgets V1 Budgets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Budgets",
        "tags": [
          "budgets"
        ]
      },
      "post": {
        "description": "Create a new budget.",
        "operationId": "create_budget_v1_budgets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBudgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Budget",
        "tags": [
          "budgets"
        ]
      }
    },
    "/v1/budgets/{budget_id}": {
      "delete": {
        "description": "Delete a budget.",
        "operationId": "delete_budget_v1_budgets__budget_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Budget",
        "tags": [
          "budgets"
        ]
      },
      "get": {
        "description": "Get details of a specific budget.",
        "operationId": "get_budget_v1_budgets__budget_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Budget",
        "tags": [
          "budgets"
        ]
      },
      "patch": {
        "description": "Update a budget.",
        "operationId": "update_budget_v1_budgets__budget_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Budget",
        "tags": [
          "budgets"
        ]
      }
    },
...
    "/v1/pricing": {
      "get": {
        "description": "List all model pricing.",
        "operationId": "list_pricing_v1_pricing_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PricingResponse"
                  },
                  "title": "Response List Pricing V1 Pricing Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pricing",
        "tags": [
          "pricing"
        ]
      },
      "post": {
        "description": "Set or update pricing for a model.",
        "operationId": "set_pricing_v1_pricing_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPricingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Pricing",
        "tags": [
          "pricing"
        ]
      }
    },
    "/v1/pricing/{model_key}": {
      "delete": {
        "description": "Delete pricing for a model.",
        "operationId": "delete_pricing_v1_pricing__model_key__delete",
        "parameters": [
          {
            "in": "path",
            "name": "model_key",
            "required": true,
            "schema": {
              "title": "Model Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Pricing",
        "tags": [
          "pricing"
        ]
      },
      "get": {
        "description": "Get pricing for a specific model.",
        "operationId": "get_pricing_v1_pricing__model_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_key",
            "required": true,
            "schema": {
              "title": "Model Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pricing",
        "tags": [
          "pricing"
        ]
      }
    },
...
  }
}